requirements.txt
requests~=2.31.0
python-telegram-bot~=13.15
urllib3~=1.26.5
environs~=10.1.0

